acq: ddd async and sync result APIs.
This replaces the previous API, which had to be driven by the
application, with an API where we directly get the result. For cases
where async or timed operation are desired, it can do either as well.

We are using C++17 variants for now. If there are better API options,
they can be investigated in the future.