Replace libpcidriver with kernel APIs.
Thanks to [1] and the initialization in [2], we found we can simply use
files under /sys/bus/pci to map physical slots to our PCI devices and to
then memory map their BARs.

This commit removes usage of the libpcidriver and its associated driver
entirely. This means it now requires something else to set permissions
in the resourcesX files appropriately so they can be accessed without
root. This also means we no longer have access to DMA requests, which on
the plus side simplifies things greatly with, given the size of our data
transfers, likely low performance impact.

We still need an equivalent dev_close() function, and, ideally, a C++
class to automatically discard resources when the device goes out of
scope. Right now we rely on the kernel doing cleanup for us when the
program exits.

[1] https://gitlab.cern.ch/tgingold/ual
[2] https://github.com/lnls-dig/fpga_pcie_driver/blob/00121b9cbec3693821eaa6e51d1e771bcdc716af/drivers/pcie/src/base.c#L359